Add workers module, separate handshake_worker into multiple fns
